scripts/requirements.txt
feedparser==6.0.11
trafilatura==1.12.5
requests==2.32.3
python-dateutil==2.9.0
jinja2==3.1.4